Removed download report button from sidebar and added in dashboard page
